autoconfig: manage windows platform
