+++ cmd/agent/common/auto_listener.go
@@ -3,12 +3,12 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2017 Datadog, Inc.
 
+// +build !docker
+
 package common
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/config"
-	"github.com/DataDog/datadog-agent/pkg/util/docker"
-	log "github.com/cihub/seelog"
 )
 
 // AutoAddListeners checks if the listener auto is selected and
@@ -17,46 +17,5 @@ import (
 // Note: auto listener isn't a real listener but a way to automatically starts other listeners
 // TODO: support more listeners (kubelet, ...)
 func AutoAddListeners(listeners []config.Listeners) []config.Listeners {
-	autoIdx := -1
-	for i, l := range listeners {
-		switch l.Name {
-		case "docker":
-			return listeners
-		case "auto":
-			autoIdx = i
-		}
-	}
-	if autoIdx == -1 {
-		return listeners
-	}
-
-	// Remove the auto listener element from the listeners slice
-	listeners = remove(listeners, autoIdx)
-
-	if isDockerRunning() == false {
-		return listeners
-	}
-
-	// Adding listeners
-	listeners = addListener(listeners, "docker")
-	log.Debugf("returning %d listeners", len(listeners))
 	return listeners
-}
-
-func addListener(listeners []config.Listeners, name string) []config.Listeners {
-	newListener := config.Listeners{Name: name}
-	log.Infof("auto adding %q listener", newListener.Name)
-	return append(listeners, newListener)
-}
-
-// isDockerRunning check if docker is running
-func isDockerRunning() bool {
-	_, err := docker.GetDockerUtil()
-	return err == nil
-}
-
-// remove the elt at the index
-func remove(s []config.Listeners, index int) []config.Listeners {
-	s[index] = s[len(s)-1]
-	return s[:len(s)-1]
 }
